deleted stuff that isnt necessary
